+++ lib/src/service/spaces/spaces_service.dart
@@ -21,6 +21,8 @@ import 'space_data.dart';
 import 'space_expansion.dart';
 import 'space_field.dart';
 import 'space_meta.dart';
+import 'space_state.dart';
+import 'topic_field.dart';
 
 abstract class SpacesService {
   /// Returns the new instance of [SpacesService].
@@ -39,6 +41,11 @@ abstract class SpacesService {
   ///
   /// - [query]: Your search term. This can be any text (including mentions and
   ///            Hashtags) present in the title of the Space.
+  ///
+  /// - [state]: Determines the type of results to return.
+  ///            This endpoint returns `all` Spaces by default. Use `live` to
+  ///            only return live Spaces or scheduled to only return
+  ///            upcoming Spaces.
   ///
   /// - [expansions]: Expansions enable you to request additional data objects
   ///                 that relate to the originally returned Space. Submit a
@@ -59,6 +66,11 @@ abstract class SpacesService {
   /// - [spaceFields]: This fields parameter enables you to select which
   ///                  specific Space fields will deliver in each returned
   ///                  Space.
+  ///
+  /// - [topicFields]: This fields parameter enables you to select which
+  ///                  specific topic metadata in each returned Space topic
+  ///                  object, if the creator of the Space set one or more
+  ///                  topics.
   ///
   /// ## Endpoint Url
   ///
@@ -88,9 +100,11 @@ abstract class SpacesService {
   /// - https://developer.twitter.com/en/docs/twitter-api/spaces/search/api-reference/get-spaces-search
   Future<TwitterResponse<List<SpaceData>, SpaceMeta>> search({
     required String query,
+    SpaceState? state,
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   });
 
   /// Returns a variety of information about a single Space specified by the
@@ -119,6 +133,11 @@ abstract class SpacesService {
   /// - [spaceFields]: This fields parameter enables you to select which
   ///                  specific Space fields will deliver in each returned
   ///                  Space.
+  ///
+  /// - [topicFields]: This fields parameter enables you to select which
+  ///                  specific topic metadata in each returned Space topic
+  ///                  object, if the creator of the Space set one or more
+  ///                  topics.
   ///
   /// ## Endpoint Url
   ///
@@ -151,6 +170,7 @@ abstract class SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   });
 
   /// Returns details about multiple Spaces. Up to 100 comma-separated Spaces
@@ -179,6 +199,11 @@ abstract class SpacesService {
   /// - [spaceFields]: This fields parameter enables you to select which
   ///                  specific Space fields will deliver in each returned
   ///                  Space.
+  ///
+  /// - [topicFields]: This fields parameter enables you to select which
+  ///                  specific topic metadata in each returned Space topic
+  ///                  object, if the creator of the Space set one or more
+  ///                  topics.
   ///
   /// ## Endpoint Url
   ///
@@ -211,6 +236,7 @@ abstract class SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   });
 
   /// Returns a list of user who purchased a ticket to the requested Space.
@@ -446,6 +472,11 @@ abstract class SpacesService {
   ///                  specific Space fields will deliver in each returned
   ///                  Space.
   ///
+  /// - [topicFields]: This fields parameter enables you to select which
+  ///                  specific topic metadata in each returned Space topic
+  ///                  object, if the creator of the Space set one or more
+  ///                  topics.
+  ///
   /// ## Endpoint Url
   ///
   /// - https://api.twitter.com/2/spaces/by/creator_ids
@@ -480,6 +511,7 @@ abstract class SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   });
 }
 
@@ -490,9 +522,11 @@ class _SpacesService extends BaseService implements SpacesService {
   @override
   Future<TwitterResponse<List<SpaceData>, SpaceMeta>> search({
     required String query,
+    SpaceState? state,
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   }) async =>
       super.buildMultiDataResponse(
         await super.get(
@@ -500,9 +534,11 @@ class _SpacesService extends BaseService implements SpacesService {
           '/2/spaces/search',
           queryParameters: {
             'query': query,
+            'state': state?.name,
             'expansions': expansions,
             'user.fields': userFields,
             'space.fields': spaceFields,
+            'topic.fields': topicFields,
           },
         ),
         dataBuilder: SpaceData.fromJson,
@@ -515,6 +551,7 @@ class _SpacesService extends BaseService implements SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   }) async =>
       super.buildResponse(
         await super.get(
@@ -524,6 +561,7 @@ class _SpacesService extends BaseService implements SpacesService {
             'expansions': expansions,
             'user.fields': userFields,
             'space.fields': spaceFields,
+            'topic.fields': topicFields,
           },
         ),
         dataBuilder: SpaceData.fromJson,
@@ -535,6 +573,7 @@ class _SpacesService extends BaseService implements SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   }) async =>
       super.buildMultiDataResponse(
         await super.get(
@@ -545,6 +584,7 @@ class _SpacesService extends BaseService implements SpacesService {
             'expansions': expansions,
             'user.fields': userFields,
             'space.fields': spaceFields,
+            'topic.fields': topicFields,
           },
         ),
         dataBuilder: SpaceData.fromJson,
@@ -609,6 +649,7 @@ class _SpacesService extends BaseService implements SpacesService {
     List<SpaceExpansion>? expansions,
     List<UserField>? userFields,
     List<SpaceField>? spaceFields,
+    List<TopicField>? topicFields,
   }) async =>
       super.buildMultiDataResponse(
         await super.get(
@@ -619,6 +660,7 @@ class _SpacesService extends BaseService implements SpacesService {
             'expansions': expansions,
             'user.fields': userFields,
             'space.fields': spaceFields,
+            'topic.fields': topicFields,
           },
         ),
         dataBuilder: SpaceData.fromJson,